+++ scooter-service/src/main/java/ar/edu/unicen/scooterservice/domain/dtos/report/NearScooterReportDTO.java
@@ -2,7 +2,7 @@ package ar.edu.unicen.scooterservice.domain.dtos.report;
 
 public record NearScooterReportDTO(
         Long scooterId,
-        float latitude,
-        float longitude
+        double latitude,
+        double longitude
 ) {
 }
